+++ cpukit/score/src/schedulercbsreleasejob.c
@@ -24,7 +24,7 @@
 void _Scheduler_CBS_Release_job(
   const Scheduler_Control *scheduler,
   Thread_Control          *the_thread,
-  uint32_t                 deadline
+  uint64_t                 deadline
 )
 {
   Scheduler_CBS_Node   *node;
