+++ cpukit/score/src/schedulerdefaultreleasejob.c
@@ -24,10 +24,10 @@
 void _Scheduler_default_Release_job(
   const Scheduler_Control *scheduler,
   Thread_Control          *the_thread,
-  uint32_t                 deadline
+  uint64_t                 deadline
 )
 {
-  ( void ) scheduler;
-  ( void ) the_thread;
-  ( void ) deadline;
+  (void) scheduler;
+  (void) the_thread;
+  (void) deadline;
 }
